requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-chess==1.1.0
websockets==15.0.1